Moved Structs out of Call Concluder
