+++ trunk-recorder/systems/system.h
@@ -5,10 +5,9 @@
 #include <boost/log/trivial.hpp>
 #include <boost/foreach.hpp>
 #include <stdio.h>
-//#include "../source.h"
-#include "p25_trunking.h"
+//#include "../source.h"ß
 #include "parser.h"
-#include "smartnet_trunking.h"
+
 
 #ifdef __GNUC__
 #pragma GCC diagnostic push
@@ -16,7 +15,6 @@
 //#pragma GCC diagnostic ignored "-Wunused-local-typedefs"
 #endif
 
-#include <lfsr/lfsr.h>
 
 #ifdef __GNUC__
 #pragma GCC diagnostic pop
@@ -39,7 +37,7 @@ enum TalkgroupDisplayFormat { talkGroupDisplayFormat_id = 0,
 	#else
   typedef std::shared_ptr<analog_recorder> analog_recorder_sptr;
   typedef std::shared_ptr<p25_recorder> p25_recorder_sptr;
-  typedef std::shared_ptr<dmr_recorder> dmr_recorder_sptr;
+  typedef std::shared_ptr<dmr_recorder> dmr_recorder_sptr;ß
 	#endif
 
 class System {
